+++ ichnaea/cache.py
@@ -67,7 +67,6 @@ class RedisClient(redis.StrictRedis):
         'fallback_cell': b'cache:fallback:cell:',
         'fallback_wifi': b'cache:fallback:wifi:',
         'leaders': b'cache:leaders:2',
-        'leaders_weekly': b'cache:leaders_weekly:2',
         'stats': b'cache:stats:3',
         'stats_regions': b'cache:stats_regions:4',
         'stats_blue_json': b'cache:stats_blue_json:2',